move iterators to their own files
